fix(freebsd): build, signed rlim_t, FIFO procsub fallback, hermetic itests
`libc::RLIM_INFINITY` and `rlimit::rlim_{cur,max}` are signed `i64`
on FreeBSD (where `rlim_t = __int64_t`) but unsigned `u64` on
Linux/macOS/OpenBSD; the build fails on FreeBSD without an
explicit cast. Cast both unconditionally to `u64` — the bit
pattern (`-1` ↔ `u64::MAX`) round-trips losslessly across the two
representations and the kernel ABI uses the same encoding.

Once the tree builds, the `/dev/fd` runtime probe (used to choose
between `/dev/fd/N` and FIFO backings for `bash_procsub`) returned
a false positive on FreeBSD: `/dev/fd` is a regular `devfs`
directory containing only static entries `0`, `1`, `2` — the live
fdescfs mount lives at `/compat/linux/dev/fd`. The previous
"directory exists" check passed, then the subshell's
`open(/dev/fd/3)` failed with `ENOENT` at runtime. Replace the
probe with positive evidence: open a fresh pipe, `stat(2)` the
pipe's read fd through `/dev/fd/N`, and accept the answer only if
that succeeds.

The FIFO fallback path also needed work for the no-consumer case
(`printf '%s' <(producer)`): the subshell's `open(FIFO,
O_WRONLY)` blocks on rendezvous with a reader that never opens.
On `/dev/fd` the kernel keeps the substitution unblocked through
the parent's pipe, but FIFO mode has no such guarantee. Send
`SIGTERM` from `cleanup_lease` when a non-blocking probe finds
the subshell still alive at cleanup time. This is a deliberate,
FIFO-only deviation from spec § 7.2 ("the parent shall not
interpose any timeout or signal of its own"); cooperating
subshells exit before cleanup runs and never see the signal, so
the corner case is observable only for benign constructs that
the spec leaves under-defined for FIFO.

Two integration tests pinned `/dev/fd/`-shaped path strings;
they now accept either implementation-defined form per
spec § 6.4. `run_interactive` was inheriting the developer's
`HOME`, so on FreeBSD the spawned `meiksh -i` ran the default
`~/.profile` which calls `fortune freebsd-tips`; the random tip
text bled into stdout and broke `fc_reexec_mode`'s word-counting
assertion. Point `HOME` at a fresh empty directory and drop
`ENV` so the helper is hermetic; the dedicated startup-file
sourcing tests already supply their own `HOME` and use
`Command::new(meiksh())` directly, so they are unaffected.

All 615 unit/integration tests and all 1858 matrix tests pass on
Linux, OpenBSD, and FreeBSD.

+++ src/sys/constants.rs
@@ -151,7 +151,12 @@ pub(crate) const RLIMIT_CPU: i32 = libc::RLIMIT_CPU as i32;
 pub(crate) const RLIMIT_AS: Option<i32> = Some(libc::RLIMIT_AS as i32);
 #[cfg(target_os = "openbsd")]
 pub(crate) const RLIMIT_AS: Option<i32> = None;
-pub(crate) const RLIM_INFINITY: u64 = libc::RLIM_INFINITY;
+// `libc::RLIM_INFINITY` is `u64` (`rlim_t`) on Linux/macOS/OpenBSD but
+// `i64` on FreeBSD, where `rlim_t = __int64_t`. Cast unconditionally to
+// `u64` so the constant has a single, portable type — the bit pattern
+// (`-1` ↔ `u64::MAX`) is identical across both representations, which
+// is exactly what the kernel ABI uses.
+pub(crate) const RLIM_INFINITY: u64 = libc::RLIM_INFINITY as u64;
 
 #[cfg(test)]
 mod tests {
